modify ManageContainers
