+++ model/IAdmin.java
@@ -1,5 +1,4 @@
-
-import java.util.ArrayList;
+package model;
 
 public interface IAdmin extends IManager {
     void addPort(); // Exclusive method for admin to add a new port
