+++ model/IUser.java
@@ -1,3 +1,5 @@
+package model;
+
 public interface IUser {
